fastapi/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
prophet
lightgbm
SQLAlchemy
pymysql
python-dotenv
